backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Paper Insight - arXiv paper fetching and summarization API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "arxiv>=2.3.1",
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
